Modified to address PR requested modifications.
